backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<1.1
Pillow>=10.0,<11.0
django-anymail[resend]==12.0

